BackEnd/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Tech Blog BackEnd"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"

dependencies = [
    "fastapi (==0.121.0)",
    "aiosqlite (==0.21.0)",
    "alembic (==1.17.2)",
    "annotated-doc (==0.0.3)",
    "annotated-types (==0.7.0)",
    "anyio (==4.11.0)",
    "argon2-cffi (==25.1.0)",
    "argon2-cffi-bindings (==25.1.0)",
    "bcrypt (==5.0.0)",
    "bleach (==6.3.0)",
    "build (==1.3.0)",
    "cachecontrol (==0.14.4)",
    "certifi (==2025.10.5)",
    "cffi (==2.0.0)",
    "charset-normalizer (==3.4.4)",
    "cleo (==2.1.0)",
    "click (==8.3.0)",
    "colorama (==0.4.6)",
    "cowsay (==6.1)",
    "crashtest (==0.4.1)",
    "cryptography (==46.0.3)",
    "deprecated (==1.3.1)",
    "distlib (==0.4.0)",
    "dnspython (==2.8.0)",
    "dotenv (==0.9.9)",
    "dulwich (==0.24.10)",
    "ecdsa (==0.19.1)",
    "email-validator (==2.3.0)",
    "fastjsonschema (==2.21.2)",
    "filelock (==3.20.0)",
    "findpython (==0.7.1)",
    "greenlet (==3.2.4)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "idna (==3.11)",
    "iniconfig (==2.3.0)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.3.0)",
    "keyring (==25.7.0)",
    "limits (==5.6.0)",
    "mako (==1.3.10)",
    "markupsafe (==3.0.3)",
    "more-itertools (==10.8.0)",
    "msgpack (==1.1.2)",
    "packaging (==25.0)",
    "passlib (==1.7.4)",
    "pbs-installer (==2025.11.20)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.5.0)",
    "pluggy (==1.6.0)",
    "poetry (==2.2.1)",
    "poetry-core (==2.2.1)",
    "pyasn1 (==0.6.1)",
    "pycparser (==2.23)",
    "pydantic (==2.12.3)",
    "pydantic-settings (==2.12.0)",
    "pydantic-core (==2.41.4)",
    "pygments (==2.19.2)",
    "pyproject-hooks (==1.2.0)",
    "pytest (==9.0.1)",
    "python-dotenv (==1.2.1)",
    "python-jose (==3.5.0)",
    "pywin32-ctypes (==0.2.3)",
    "rapidfuzz (==3.14.3)",
    "requests (==2.32.5)",
    "requests-toolbelt (==1.0.0)",
    "rsa (==4.9.1)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "slowapi (==0.1.9)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.44)",
    "starlette (==0.49.3)",
    "structlog (==25.5.0)",
    "tomlkit (==0.13.3)",
    "trove-classifiers (==2025.11.14.15)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.15.0)",
    "urllib3 (==2.5.0)",
    "uvicorn (==0.38.0)",
    "virtualenv (==20.35.3)",
    "webencodings (==0.5.1)",
    "wrapt (==2.0.1)",
    "zstandard (==0.25.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
